Implement popSceneWithTransition method
Pops out a scene from the queue with transition
This scene will replace the running one
The running scene will be deleted. If there are no more scenes in the stack the execution is terminated
ONLY call it if there is a running scene.
Reference: https://discuss.cocos2d-x.org/t/tip-implement-popscene-with-transition/3647